requirements.txt
python-dotenv==1.0.0
langchain-groq
langchain
langgraph
